example: simplify major version updates
This commit is an example of a change to be rolled out with ci-mgmt.

Providers are free to modify their major version without requiring many code
changes beyond changing the code in `./provider`:

1. disabling examples tests (can be done with PR changing provider code)
2. ci-mgmt change to bump the major version
3. ci-mgmt PR against repo updating major version in CI files
4. re-enabling examples

To accomplish this, files managed via ci-mgmt no longer require using the major
version number of the provider in any scripts. This major version appears in:

* Makefile: no longer required as go commands now use paths to build and
  modifying ldflags version argument
* goreleaser: no longer required by modifying version ldflags argument
* GitHub Actions: no longer required as an environment variable

A change to pulumi/pkg ProgramTest is needed to modify `Dependencies` to remove
the need to disable and re-enable examples:
- https://github.com/pulumi/pulumi/pull/12004